code/requirements.txt
langchain==0.1.14
langchain-openai==0.0.8
langchain-community==0.0.30
python-dotenv==1.0.0
PyPDF2==3.0.1
tiktoken==0.5.2
chromadb==0.4.22
pypdf==3.17.1
openai>=1.6.1,<2.0.0 